requirements.txt
fastapi==0.82.0
gunicorn==20.0.2
uvicorn==0.18.3
pytest==7.1.3
requests==2.28.1
pre-commit==2.20.0
python-dotenv==0.21.0
aiofiles==22.1.0
Pillow==9.2.0
python-multipart==0.0.5
pytesseract==0.3.10